chore: refactored components to file, ui: improved settings login view
